Add testing config Add first test Extract trophy scanner functionality to module Refactor trophy scanner functionality
