+++ .gitignore
@@ -8,3 +8,4 @@ config.ini
 __pycache__/
 scratch.py
 .idea/
+.coverage
