+++ read_izgc_trophies.py
@@ -1,16 +1,16 @@
 import sys
 
 from dispatcher import Dispatcher
-from thread_reader import Thread
+from trophy_scanner import report_new_trophies, IZGCThread
 
 dispatcher = Dispatcher()
 dispatcher.login(required=False)
-
-club_thread = Thread(dispatcher, dispatcher.izgc_thread_id())
+club_thread = IZGCThread(dispatcher=dispatcher)
 
 # Accept command line argument to read all pages
 if "--all-pages" in sys.argv:
     club_thread.page_number = 1
     club_thread.last_post = 0
 
-club_thread.trophy_scan()
+imp_trophies = club_thread.trophy_scan()
+report_new_trophies(imp_trophies)
