+++ setup.cfg
@@ -1,0 +1,6 @@
+[coverage:run]
+branch = True
+
+[coverage:report]
+show_missing = True
+skip_covered = True
