+++ trophy_scanner.py
@@ -1,0 +1,115 @@
+"""Functionality for parsing IZGC thread for new trophies"""
+
+import json
+import re
+
+from bs4 import BeautifulSoup
+
+from thread_reader import Thread
+
+
+def report_new_trophies(imp_trophies):
+    if not imp_trophies:
+        return print("No new trophies found.")
+
+    timestamp_file = "trophy_timestamps.json"
+    try:
+        with open(timestamp_file, "r") as file:
+            trophy_log = json.load(file)
+    except FileNotFoundError:
+        trophy_log = {}
+
+    print("\n******** NEW TROPHIES ********")
+    for imp, trophies in imp_trophies.items():
+        if imp not in trophy_log:
+            new_member_string = " (New Club member!)"
+            trophy_log[imp] = trophies
+        else:
+            new_member_string = ""
+            update_trophy_dict(trophy_log[imp], trophies)
+
+        print(f"{imp}{new_member_string}:")
+        for game, trophy in trophies.items():
+            for name, timestamp in trophy.items():
+                print(f"[{game}] {name} -- posted {timestamp}")
+
+    with open(timestamp_file, "w", encoding="utf-8") as file:
+        json.dump(trophy_log, file, indent=2)
+
+
+def get_izgc_master_trophy_dict(session):
+    # This should probably be in a config file, but I'm not putting it
+    # there just so it's a little less visible to casual perusal.
+    izgc_trophy_list_url = "https://impzone.club/alltrophies.html"
+    master_trophy_dict = {}
+
+    # Could modularize this, but it's YAGNI for now
+    response = session.get(izgc_trophy_list_url)
+    soup = BeautifulSoup(response.text, "html.parser")
+
+    trophies = soup.find_all("div", "item-trophy tooltip")
+    for trophy in trophies:
+        try:
+            trophy_info_strings = tuple(trophy.stripped_strings)
+            game = trophy_info_strings[1]
+            name = trophy_info_strings[0]
+            trophy_data = {
+                "full_name": f"[{game}] {name}",
+                "game": game,
+                "name": name,
+                "system_year": trophy_info_strings[2]
+            }
+            master_trophy_dict[trophy["imgur_id"]] = trophy_data
+        except KeyError:
+            # No imgur id present, so we don't scan for it
+            continue
+
+    return master_trophy_dict
+
+
+def update_trophy_dict(existing_trophies, new_trophies):
+    for game, game_trophies in new_trophies.items():
+        if game not in existing_trophies:
+            existing_trophies[game] = game_trophies
+        else:
+            existing_trophies[game].update(game_trophies)
+
+
+class IZGCThread(Thread):
+    """Thread with additional functionality for trophy scanning"""
+    def __init__(self, *, dispatcher):
+        self.dispatcher = dispatcher
+        super().__init__(
+            dispatcher=dispatcher,
+            thread_id=dispatcher.config["DEFAULT"]["izgc_thread_id"]
+        )
+        self.eligible_trophies = get_izgc_master_trophy_dict(
+            self.dispatcher.session)
+
+    def trophy_scan(self):
+        imp_trophies = {}
+        post_list = self.new_posts()
+
+        for post in post_list:
+            post.remove_quotes()
+            post_trophies = self.get_post_trophies(post)
+            if post_trophies:
+                if post.username not in imp_trophies:
+                    imp_trophies[post.username] = post_trophies
+                else:
+                    update_trophy_dict(
+                        imp_trophies[post.username], post_trophies)
+
+        return imp_trophies
+
+    def get_post_trophies(self, post):
+        earned_trophies = {}
+        images = post.image_urls()
+        for image in images:
+            for trophy_id, trophy_data in self.eligible_trophies.items():
+                if re.search(f"i.imgur.com/{trophy_id}", image):
+                    new_trophy = {trophy_data["game"]: {
+                        trophy_data["name"]: post.timestamp}}
+                    update_trophy_dict(earned_trophies, new_trophy)
+
+        return earned_trophies
